guided_vision_vllm-updated/requirements_server.txt
fastapi
uvicorn[standard]
pydantic
numpy
opencv-python
torch
ultralytics
python-multipart
PyYAML
